videos/serializers.py
from rest_framework import serializers
from .models import Video, Category

class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = ['id', 'name']

class VideoSerializer(serializers.ModelSerializer):
    category_name = serializers.ReadOnlyField(source='category.name')
    
    class Meta:
        model = Video
        fields = ['id', 'title', 'youtube_id', 'embed_url', 'thumbnail_url', 'category_name']
