requirements.txt
django
gunicorn
whitenoise
python-dotenv
psycopg[binary]
dj-database-url
djangorestframework
celery
